astrology-bot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic>=2.7.0
jieba==0.42.1
scikit-learn>=1.3.0
aiofiles==24.1.0
